// libs/designer-v2/src/lib/core/actions/bjsworkflow/add.d.ts
import type { NodeOperation, OperationMetadataState } from '../../state/operation/operationMetadataSlice';
import type { RelationshipIds } from '../../state/panel/panelTypes';
import { type NodesMetadata, type WorkflowState } from '../../state/workflow/workflowInterfaces';
import type { AppDispatch, RootState } from '../../store';
import type { NodeDataWithOperationMetadata } from './operationdeserializer';
import type { Settings } from './settings';
import type { Connection, Connector, DiscoveryOperation, DiscoveryResultTypes, OperationManifest } from '@microsoft/logic-apps-shared';
import type { Dispatch } from '@reduxjs/toolkit';
type AddOperationPayload = {
    operation: DiscoveryOperation<DiscoveryResultTypes> | undefined;
    relationshipIds: RelationshipIds;
    nodeId: string;
    isParallelBranch?: boolean;
    isTrigger?: boolean;
    presetParameterValues?: Record<string, any>;
    actionMetadata?: Record<string, any>;
    isAddingHandoff?: boolean;
    isAddingMcpServer?: boolean;
    connectionId?: string;
};
export declare const addOperation: import("@reduxjs/toolkit").AsyncThunk<void, AddOperationPayload, {
    state?: unknown;
    dispatch?: Dispatch<import("@reduxjs/toolkit").AnyAction> | undefined;
    extra?: unknown;
    rejectValue?: unknown;
    serializedErrorType?: unknown;
    pendingMeta?: unknown;
    fulfilledMeta?: unknown;
    rejectedMeta?: unknown;
}>;
export declare const addDefaultSecureSettings: (settings: Settings, isSecureByDefault: boolean) => Settings;
export declare const initializeOperationDetails: (nodeId: string, operationInfo: NodeOperation, getState: () => RootState, dispatch: Dispatch, presetParameterValues?: Record<string, any>, actionMetadata?: Record<string, any>, openPanel?: boolean, connectionId?: string) => Promise<void>;
export declare const initializeSubgraphFromManifest: (id: string, manifest: OperationManifest, dispatch: Dispatch) => Promise<void>;
export declare const trySetDefaultConnectionForNode: (nodeId: string, connector: Connector, dispatch: AppDispatch, isConnectionRequired: boolean, connectionFilterFunc?: ((connection: Connection) => boolean) | undefined, preferredConnectionId?: string) => Promise<void>;
export declare const addTokensAndVariables: (nodeId: string, operationType: string, nodeData: NodeDataWithOperationMetadata, state: RootState, dispatch: Dispatch) => void;
export declare const getTriggerNodeManifest: (workflowState: WorkflowState, operations: OperationMetadataState) => Promise<OperationManifest | undefined>;
export declare const getNonDuplicateNodeId: (nodesMetadata: NodesMetadata, actionId: string, idReplacements?: Record<string, string>) => string;
export declare const getNonDuplicateId: (existingActionNames: Record<string, string>, actionId: string) => string;
export declare const tryGetMostRecentlyUsedConnectionId: (connectorId: string, allConnections: Connection[]) => Promise<Connection | undefined>;
export {};